refactor: move story-helpers into styleguide/
Simplifies the import path
