+++ styleguide/components/asset-hyperlink/asset-hyperlink.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Preview, Source } from '@storybook/addon-docs/blocks';
-import { translate } from '../../../scripts/story-helpers';
+import { translate } from '../../story-helpers';
 
 import templateExample from './_example.html.haml';
