+++ styleguide/components/breadcrumbs/breadcrumbs.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Preview, Source } from '@storybook/addon-docs/blocks';
-import { translate } from '../../../scripts/story-helpers';
+import { translate } from '../../story-helpers';
 
 import templateDefault from './_default.html.haml';
 import templateNoCollapse from './_no-collapse.html.haml';
