+++ styleguide/components/footer/footer.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Source, Story, Preview } from '@storybook/addon-docs/blocks';
-import { translate } from '../../../scripts/story-helpers';
+import { translate } from '../../story-helpers';
 
 import templateDefault from './_default.html.haml';
 import templateMinimal from './_minimal.html.haml';
