+++ styleguide/components/header/header.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Source, Story, Preview } from '@storybook/addon-docs/blocks';
-import { translate } from '../../../scripts/story-helpers';
+import { translate } from '../../story-helpers';
 
 import initHeader from '../../../src/js/header';
 import templateDefault from './_default.html.haml';
