+++ styleguide/forms/input/input.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Preview, Source } from '@storybook/addon-docs/blocks';
-import { translate } from '../../../scripts/story-helpers';
+import { translate } from '../../story-helpers';
 
 import templateInput from './_input.html.haml';
 import templateInputWithHint from './_input-with-hint.html.haml';
