+++ styleguide/forms/radio_group/radio_group.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Preview, Source } from '@storybook/addon-docs/blocks';
-import { translate } from '../../../scripts/story-helpers';
+import { translate } from '../../story-helpers';
 
 import templateRadioGroupWithHint from './_radio_group-with-hint.html.haml';
 import templateRadioGroupOptional from './_radio_group-optional.html.haml';
