+++ styleguide/forms/success-message/success-message.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Preview } from '@storybook/addon-docs/blocks';
-import { translate } from '../../../scripts/story-helpers';
+import { translate } from '../../story-helpers';
 
 <Meta title="Forms/Success message" />
